+++ CHANGELOG.md
@@ -13,3 +13,44 @@
 ## [0.0.4] - 2019.07.09
 
 * update common sp.
+
+## [0.0.5] - 2019.07.11
+
+* add common arc clip.
+
+## [0.0.6] - 2019.07.11
+
+* update common arc clip.
+
+## [0.0.7] - 2019.07.11
+
+* update common arc clip.
+
+## [0.0.8] - 2019.07.11
+
+* add common load more view.
+
+## [0.0.9] - 2019.07.11
+
+* update common load more view.
+
+## [0.1.0] - 2019.07.11
+
+* add common web view.
+
+## [0.1.1] - 2019.07.11
+
+* update common web view.
+
+## [0.1.2] - 2019.07.11
+
+* add common slide view.
+
+## [0.1.3] - 2019.07.11
+
+* update common slide view.
+* add common launch.
+
+## [0.1.4] - 2019.07.11
+
+* update common web view.